+++ Connectivity_4D/src/lagrange/utils/Fmath.java
@@ -66,7 +66,7 @@ public class Fmath{
 
         // Log to base 10 of a float number
         public static float log10(float a){
-            return (float) (Math.log((double)a)/Math.log(10.0D));
+            return (float) (Math.log(a)/Math.log(10.0D));
         }
 
         // Base 10 antilog of a double
@@ -76,7 +76,7 @@ public class Fmath{
 
         // Base 10 antilog of a float
         public static float antilog10(float x){
-            return (float)Math.pow(10.0D, (double)x);
+            return (float)Math.pow(10.0D, x);
         }
 
         // Log to base e of a double number
@@ -86,7 +86,7 @@ public class Fmath{
 
         // Log to base e of a float number
         public static float log(float a){
-            return (float)Math.log((double)a);
+            return (float)Math.log(a);
         }
 
         // Base e antilog of a double
@@ -96,7 +96,7 @@ public class Fmath{
 
         // Base e antilog of a float
         public static float antilog(float x){
-            return (float)Math.exp((double)x);
+            return (float)Math.exp(x);
         }
 
         // Log to base 2 of a double number
@@ -106,7 +106,7 @@ public class Fmath{
 
         // Log to base 2 of a float number
         public static float log2(float a){
-            return (float) (Math.log((double)a)/Math.log(2.0D));
+            return (float) (Math.log(a)/Math.log(2.0D));
         }
 
         // Base 2 antilog of a double
@@ -116,7 +116,7 @@ public class Fmath{
 
         // Base 2 antilog of a float
         public static float antilog2(float x){
-            return (float)Math.pow(2.0D, (double)x);
+            return (float)Math.pow(2.0D, x);
         }
 
         // Log to base b of a double number and double base
@@ -126,17 +126,17 @@ public class Fmath{
 
         // Log to base b of a double number and int base
         public static double log10(double a, int b){
-            return Math.log(a)/Math.log((double)b);
+            return Math.log(a)/Math.log(b);
         }
 
         // Log to base b of a float number and flaot base
         public static float log10(float a, float b){
-            return (float) (Math.log((double)a)/Math.log((double)b));
+            return (float) (Math.log(a)/Math.log(b));
         }
 
         // Log to base b of a float number and int base
         public static float log10(float a, int b){
-            return (float) (Math.log((double)a)/Math.log((double)b));
+            return (float) (Math.log(a)/Math.log(b));
         }
 
         // SQUARES
@@ -1008,7 +1008,7 @@ public class Fmath{
         // x and y are int, percentage provided as double
         public static boolean isEqualWithinPerCent(int x, int y, double perCent){
             boolean test=false;
-            double limit = Math.abs((double)(x+y)*perCent/200.0D);
+            double limit = Math.abs((x+y)*perCent/200.0D);
             if(Math.abs(x-y)<=limit)test=true;
             return test;
         }
@@ -1240,10 +1240,10 @@ public class Fmath{
         // x is float but must hold an integer value
         public static boolean isEven(float x){
             double y=Math.floor(x);
-            if(((double)x - y)!= 0.0D)throw new IllegalArgumentException("the argument is not an integer");
+            if((x - y)!= 0.0D)throw new IllegalArgumentException("the argument is not an integer");
             boolean test=false;
             y=Math.floor(x/2.0F);
-            if(((double)(x/2.0F)-y) == 0.0D)test=true;
+            if((x/2.0F-y) == 0.0D)test=true;
             return test;
         }
 
@@ -1271,10 +1271,10 @@ public class Fmath{
         // x is float but must hold an integer value
         public static boolean isOdd(float x){
             double y=Math.floor(x);
-            if(((double)x - y)!= 0.0D)throw new IllegalArgumentException("the argument is not an integer");
+            if((x - y)!= 0.0D)throw new IllegalArgumentException("the argument is not an integer");
             boolean test=true;
             y=Math.floor(x/2.0F);
-            if(((double)(x/2.0F)-y) == 0.0D)test=false;
+            if((x/2.0F-y) == 0.0D)test=false;
             return test;
         }
 
@@ -1612,7 +1612,7 @@ public class Fmath{
             int n = aa.length;
             double[] bb = new double[n];
             for(int i=0; i<n; i++){
-               bb[i] = (double)aa[i]*constant;
+               bb[i] = aa[i]*constant;
             }
             return bb;
         }
@@ -1621,7 +1621,7 @@ public class Fmath{
             int n = aa.length;
             double[] bb = new double[n];
             for(int i=0; i<n; i++){
-               bb[i] = aa[i]*(double)constant;
+               bb[i] = aa[i]*constant;
             }
             return bb;
         }
@@ -1631,7 +1631,7 @@ public class Fmath{
             int n = aa.length;
             double[] bb = new double[n];
             for(int i=0; i<n; i++){
-               bb[i] = (double)(aa[i]*constant);
+               bb[i] = aa[i]*constant;
             }
             return bb;
         }
@@ -2657,7 +2657,7 @@ public class Fmath{
             int n = aa.length;
             double[] bb = new double[n];
             for(int i=0; i<n; i++){
-               bb[i] = (double)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2667,7 +2667,7 @@ public class Fmath{
             int n = aa.length;
             double[] bb = new double[n];
             for(int i=0; i<n; i++){
-               bb[i] = (double)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2677,7 +2677,7 @@ public class Fmath{
             int n = aa.length;
             float[] bb = new float[n];
             for(int i=0; i<n; i++){
-               bb[i] = (float)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2687,7 +2687,7 @@ public class Fmath{
             int n = aa.length;
             long[] bb = new long[n];
             for(int i=0; i<n; i++){
-               bb[i] = (long)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2698,7 +2698,7 @@ public class Fmath{
             int n = aa.length;
             double[] bb = new double[n];
             for(int i=0; i<n; i++){
-               bb[i] = (double)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2709,7 +2709,7 @@ public class Fmath{
             int n = aa.length;
             float[] bb = new float[n];
             for(int i=0; i<n; i++){
-               bb[i] = (float)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2719,7 +2719,7 @@ public class Fmath{
             int n = aa.length;
             double[] bb = new double[n];
             for(int i=0; i<n; i++){
-               bb[i] = (double)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2729,7 +2729,7 @@ public class Fmath{
             int n = aa.length;
             float[] bb = new float[n];
             for(int i=0; i<n; i++){
-               bb[i] = (float)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2739,7 +2739,7 @@ public class Fmath{
             int n = aa.length;
             long[] bb = new long[n];
             for(int i=0; i<n; i++){
-               bb[i] = (long)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2749,7 +2749,7 @@ public class Fmath{
             int n = aa.length;
             int[] bb = new int[n];
             for(int i=0; i<n; i++){
-               bb[i] = (int)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2759,7 +2759,7 @@ public class Fmath{
             int n = aa.length;
             double[] bb = new double[n];
             for(int i=0; i<n; i++){
-               bb[i] = (int)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2769,7 +2769,7 @@ public class Fmath{
             int n = aa.length;
             float[] bb = new float[n];
             for(int i=0; i<n; i++){
-               bb[i] = (float)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2779,7 +2779,7 @@ public class Fmath{
             int n = aa.length;
             long[] bb = new long[n];
             for(int i=0; i<n; i++){
-               bb[i] = (long)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2789,7 +2789,7 @@ public class Fmath{
             int n = aa.length;
             int[] bb = new int[n];
             for(int i=0; i<n; i++){
-               bb[i] = (int)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
@@ -2799,7 +2799,7 @@ public class Fmath{
             int n = aa.length;
             short[] bb = new short[n];
             for(int i=0; i<n; i++){
-               bb[i] = (short)aa[i];
+               bb[i] = aa[i];
             }
             return bb;
         }
